refactor: split agents/ into training/, data_pipeline/, and slim agents/
Split the overloaded embodichain/agents/ package into three coherent
top-level modules:

- training/rl/ : RL training subsystem (moved from agents/rl/)
- data_pipeline/{datasets,engine}/ : online streaming data infra
  (moved from agents/{datasets,engine}/)
- agents/ : slimmed to hierarchy/ + mllm/ (LLM-driven agents and their
  prompt scaffolding)

Pure structural move with import rewrites; no behavior change. All
consumer imports in embodichain/, scripts/, tests/, examples/, and docs/
are updated to the new module paths, including internal absolute imports
inside the moved rl/ subtree. Tests and examples are relocated to mirror
the new structure (tests/training/, tests/data_pipeline/,
examples/data_pipeline/). setup.py is unchanged (find_packages
auto-discovers the new packages).

Verified: package imports cleanly; pytest tests/training/
tests/data_pipeline/ -> 41 passed, 0 failed (incl. distributed path);
Sphinx docs build succeeds; black clean.
