+++ docs/source/api_reference/embodichain/embodichain.data_pipeline.datasets.rst
@@ -1,0 +1,30 @@
+embodichain.data_pipeline.datasets
+==================================
+
+.. automodule:: embodichain.data_pipeline.datasets
+
+Overview
+--------
+
+Datasets and samplers for online streaming training from live simulation.
+
+   .. rubric:: Functions
+
+   .. autosummary::
+
+      OnlineDataset
+
+.. automodule:: embodichain.data_pipeline.datasets
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+.. automodule:: embodichain.data_pipeline.datasets.online_data
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+.. automodule:: embodichain.data_pipeline.datasets.sampler
+   :members:
+   :undoc-members:
+   :show-inheritance:
