+++ docs/source/api_reference/embodichain/embodichain.data_pipeline.engine.rst
@@ -1,0 +1,26 @@
+embodichain.data_pipeline.engine
+================================
+
+.. automodule:: embodichain.data_pipeline.engine
+
+Overview
+--------
+
+Online data streaming engine: a process-safe shared buffer for trajectory data.
+
+   .. rubric:: Functions
+
+   .. autosummary::
+
+      OnlineDataEngine
+      OnlineDataEngineCfg
+
+.. automodule:: embodichain.data_pipeline.engine
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+.. automodule:: embodichain.data_pipeline.engine.data
+   :members:
+   :undoc-members:
+   :show-inheritance:
