+++ docs/source/api_reference/embodichain/embodichain.data_pipeline.rst
@@ -1,0 +1,36 @@
+embodichain.data_pipeline
+=========================
+
+.. automodule:: embodichain.data_pipeline
+
+   .. rubric:: Submodules
+
+   .. autosummary::
+
+      datasets
+      engine
+
+Datasets
+--------
+
+.. automodule:: embodichain.data_pipeline.datasets
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+   .. autosummary::
+
+      online_data
+      sampler
+
+Online Data Engine
+------------------
+
+.. automodule:: embodichain.data_pipeline.engine
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+   .. autosummary::
+
+      data
