+++ docs/source/api_reference/embodichain/embodichain.training.rl.algo.rst
@@ -1,7 +1,7 @@
-﻿embodichain.agents.rl.algo
-==========================
+embodichain.training.rl.algo
+============================
 
-.. automodule:: embodichain.agents.rl.algo
+.. automodule:: embodichain.training.rl.algo
 
 Overview
 --------
@@ -16,7 +16,7 @@ Algorithm registry and algorithm-construction helpers for RL training.
       build_algo
       get_registered_algo_names
 
-.. automodule:: embodichain.agents.rl.algo
+.. automodule:: embodichain.training.rl.algo
    :members:
    :undoc-members:
    :show-inheritance:
