+++ docs/source/api_reference/embodichain/embodichain.training.rst
@@ -1,0 +1,28 @@
+embodichain.training
+====================
+
+.. automodule:: embodichain.training
+
+   .. rubric:: Submodules
+
+   .. autosummary::
+
+      rl
+
+Reinforcement Learning
+----------------------
+
+.. automodule:: embodichain.training.rl
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+   .. autosummary::
+      :toctree: .
+
+      algo
+      buffer
+      collector
+      models
+      train
+      utils
