+++ embodichain/training/__init__.py
@@ -14,7 +14,6 @@
 # limitations under the License.
 # ----------------------------------------------------------------------------
 
-from . import hierarchy
-from . import mllm
+from . import rl
 
-__all__ = ["hierarchy", "mllm"]
+__all__ = ["rl"]
