+++ embodichain/training/rl/algo/ppo.py
@@ -19,8 +19,8 @@ from typing import Dict
 import torch
 from tensordict import TensorDict
 
-from embodichain.agents.rl.buffer import iterate_minibatches, transition_view
-from embodichain.agents.rl.utils import AlgorithmCfg
+from embodichain.training.rl.buffer import iterate_minibatches, transition_view
+from embodichain.training.rl.utils import AlgorithmCfg
 from embodichain.utils import configclass
 from .common import compute_gae
 from .base import BaseAlgorithm
